# Conflicts:
#	WebAPIGroup2/Service/Implement/UpLoadService.cs

+++ WebAPIGroup2/Models/DTO/AddTemplateDTO.cs
@@ -10,7 +10,7 @@ namespace WebAPIGroup2.Models.DTO
 
         public string? Name { get; set; }
 
-        public decimal? PricePlus { get; set; }
+        public float? PricePlus { get; set; }
 
         public bool? Status { get; set; }
